Moved SeedDatabase class to Seed folder
