+++ Trappist/src/Promact.Trappist.Repository/Questions/ICodeSnippetQuestionRepository.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace Promact.Trappist.Repository.Questions
+{
+    interface ICodeSnippetQuestionRepository
+    {
+    }
+}
